Fix namespaces for nuget deploy api
