+++ InventoryService.Messages/InventoryService.Messages.csproj
@@ -42,6 +42,7 @@
   <ItemGroup>
     <Compile Include="ErrorType.cs" />
     <Compile Include="InventoryServiceErrorMessageGenerator.cs" />
+    <Compile Include="Models\IRealTimeInventory.cs" />
     <Compile Include="Models\RealTimeInventory.cs" />
     <Compile Include="RealTimeInventoryException.cs" />
     <Compile Include="Request\GetInventoryMessage.cs" />
