+++ InventoryService.Messages/Models/IRealTimeInventory.cs
@@ -1,0 +1,10 @@
+﻿namespace InventoryService.Messages.Models
+{
+    public interface IRealTimeInventory
+    {
+        int Reserved { get; }
+        int Quantity { get; }
+        int Holds { get; }
+        string ProductId { get; }
+    }
+}
